+++ walkers/src/map.rs
@@ -3,39 +3,13 @@ use egui::{
 };
 
 use crate::{
-    MapMemory, Plugin, Position, Projector, Tiles, center::Center, position::AdjustedPosition,
-    tiles::draw_tiles,
+    MapMemory, Options, Plugin, Position, Projector, Tiles, center::Center,
+    position::AdjustedPosition, tiles::draw_tiles,
 };
 
 struct Layer<'a> {
     tiles: &'a mut dyn Tiles,
     transparency: f32,
-}
-
-struct Options {
-    zoom_gesture_enabled: bool,
-    drag_pan_buttons: DragPanButtons,
-    zoom_speed: f64,
-    double_click_to_zoom: bool,
-    double_click_to_zoom_out: bool,
-    zoom_with_ctrl: bool,
-    panning: bool,
-    pull_to_my_position_threshold: f32,
-}
-
-impl Default for Options {
-    fn default() -> Self {
-        Self {
-            zoom_gesture_enabled: true,
-            drag_pan_buttons: DragPanButtons::PRIMARY,
-            zoom_speed: 2.0,
-            double_click_to_zoom: false,
-            double_click_to_zoom_out: false,
-            zoom_with_ctrl: true,
-            panning: true,
-            pull_to_my_position_threshold: 0.0,
-        }
-    }
 }
 
 /// The actual map widget. Instances are to be created on each frame, as all necessary state is
